refactor(playground): split code editor templates
