+++ packages/agent-playground/src/components/playground/code-editor-templates/types.ts
@@ -1,0 +1,13 @@
+export interface IExampleTemplate {
+  name: string;
+  description: string;
+  code: string;
+}
+
+export interface IExampleTemplates {
+  basic: IExampleTemplate;
+  tools: IExampleTemplate;
+  streaming: IExampleTemplate;
+  multiProvider: IExampleTemplate;
+  plugins: IExampleTemplate;
+}
